requirements.txt
streamlit==1.31.0
openai==1.12.0
chromadb==0.4.22
python-dotenv==1.0.0
docling==1.0.0
pypdf==4.0.0
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
reportlab==4.0.9